Crate mas_oidc_client

An OpenID Connect client library for the Matrix specification.

This is part of the Matrix Authentication Service project.

Scope

The scope of this crate is to support OIDC features required by the Matrix specification according to MSC3861 and its sub-proposals.

As such, it is compatible with the OpenID Connect 1.0 specification, but also enforces Matrix-specific requirements or adds compatibility with new OAuth 2.0 features.

OpenID Connect and OAuth 2.0 Features

Grant Types:

Authorization Code

Client Credentials

Device Code (TBD)

User Info

Token:

Refresh Token

Introspection

Revocation

Dynamic Client Registration

PKCE

Pushed Authorization Requests

Matrix features

Client registration

Login

Matrix API Scopes

Logout

Modules

error

The error types used in this crate.

http_service

Reexports of traits to implement to provide a custom HTTP service for Client.

jose

requests

Methods to interact with OpenID Connect and OAuth2.0 endpoints.

types

OAuth 2.0 and OpenID Connect types.