Struct mas_oidc_client::jose::claims::Claim

pub struct Claim<T, V = ()> { /* private fields */ }

Implementations

impl<T, V> Claim<T, V>
where V: Validator<T>,

pub const fn new(claim: &'static str) -> Claim<T, V>

pub fn insert<I>( &self, claims: &mut HashMap<String, Value>, value: I ) -> Result<(), ClaimError>
where I: Into<T>, T: Serialize,

Insert a claim into the given claims map.

Errors

Returns an error if the value failed to serialize.

pub fn extract_required( &self, claims: &mut HashMap<String, Value> ) -> Result<T, ClaimError>
where T: DeserializeOwned, V: Default, <V as Validator<T>>::Error: Error + Send + Sync + 'static,

Extract a claim from the given claims map.

Errors

Returns an error if the value failed to deserialize, if its value is invalid or if the claim is missing.

pub fn extract_required_with_options<I>( &self, claims: &mut HashMap<String, Value>, validator: I ) -> Result<T, ClaimError>
where T: DeserializeOwned, I: Into<V>, <V as Validator<T>>::Error: Error + Send + Sync + 'static,

Extract a claim from the given claims map, with the given options.

Errors

Returns an error if the value failed to deserialize, if its value is invalid or if the claim is missing.

pub fn extract_optional( &self, claims: &mut HashMap<String, Value> ) -> Result<Option<T>, ClaimError>
where T: DeserializeOwned, V: Default, <V as Validator<T>>::Error: Error + Send + Sync + 'static,

Extract a claim from the given claims map, if it exists.

Errors

Returns an error if the value failed to deserialize or if its value is invalid.

pub fn extract_optional_with_options<I>( &self, claims: &mut HashMap<String, Value>, validator: I ) -> Result<Option<T>, ClaimError>
where T: DeserializeOwned, I: Into<V>, <V as Validator<T>>::Error: Error + Send + Sync + 'static,

Extract a claim from the given claims map, if it exists, with the given options.

Errors

Returns an error if the value failed to deserialize or if its value is invalid.