Enum mas_oidc_client::jose::jwa::SymmetricKey

#[non_exhaustive]
pub enum SymmetricKey {
    Hs256(Hmac<CoreWrapper<CtVariableCoreWrapper<Sha256VarCore, UInt<UInt<UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B0>, B0>, B0>, OidSha256>>>),
    Hs384(Hmac<CoreWrapper<CtVariableCoreWrapper<Sha512VarCore, UInt<UInt<UInt<UInt<UInt<UInt<UTerm, B1>, B1>, B0>, B0>, B0>, B0>, OidSha384>>>),
    Hs512(Hmac<CoreWrapper<CtVariableCoreWrapper<Sha512VarCore, UInt<UInt<UInt<UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B0>, B0>, B0>, B0>, OidSha512>>>),
}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Hs256(Hmac<CoreWrapper<CtVariableCoreWrapper<Sha256VarCore, UInt<UInt<UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B0>, B0>, B0>, OidSha256>>>)

Hs384(Hmac<CoreWrapper<CtVariableCoreWrapper<Sha512VarCore, UInt<UInt<UInt<UInt<UInt<UInt<UTerm, B1>, B1>, B0>, B0>, B0>, B0>, OidSha384>>>)

Hs512(Hmac<CoreWrapper<CtVariableCoreWrapper<Sha512VarCore, UInt<UInt<UInt<UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B0>, B0>, B0>, B0>, OidSha512>>>)

Implementations

impl SymmetricKey

pub fn new_for_alg( key: Vec<u8>, alg: &JsonWebSignatureAlg ) -> Result<SymmetricKey, InvalidAlgorithm>

Create a new symmetric key for the given algorithm with the given key.

Errors

Returns an error if the algorithm is not supported.

pub const fn hs256(key: Vec<u8>) -> SymmetricKey

Create a new symmetric key using the HS256 algorithm with the given key.

pub const fn hs384(key: Vec<u8>) -> SymmetricKey

Create a new symmetric key using the HS384 algorithm with the given key.

pub const fn hs512(key: Vec<u8>) -> SymmetricKey

Create a new symmetric key using the HS512 algorithm with the given key.