Struct mas_oidc_client::types::oidc::RpInitiatedLogoutRequest

pub struct RpInitiatedLogoutRequest {
    pub id_token_hint: Option<String>,
    pub logout_hint: Option<String>,
    pub client_id: Option<String>,
    pub post_logout_redirect_uri: Option<Url>,
    pub state: Option<String>,
    pub ui_locales: Option<Vec<LanguageTag>>,
}

The body of a request to the RP-Initiated Logout Endpoint.

Fields
id_token_hint: Option<String>

ID Token previously issued by the OP to the RP.

Recommended, used as a hint about the End-User’s current authenticated session with the Client.

logout_hint: Option<String>

Hint to the Authorization Server about the End-User that is logging out.

The value and meaning of this parameter is left up to the OP’s discretion. For instance, the value might contain an email address, phone number, username, or session identifier pertaining to the RP’s session with the OP for the End-User.

client_id: Option<String>

OAuth 2.0 Client Identifier valid at the Authorization Server.

The most common use case for this parameter is to specify the Client Identifier when post_logout_redirect_uri is used but id_token_hint is not. Another use is for symmetrically encrypted ID Tokens used as id_token_hint values that require the Client Identifier to be specified by other means, so that the ID Tokens can be decrypted by the OP.

post_logout_redirect_uri: Option<Url>

URI to which the RP is requesting that the End-User’s User Agent be redirected after a logout has been performed.

The value MUST have been previously registered with the OP, using the post_logout_redirect_uris registration parameter.

state: Option<String>

Opaque value used by the RP to maintain state between the logout request and the callback to the endpoint specified by the post_logout_redirect_uri parameter.

ui_locales: Option<Vec<LanguageTag>>

End-User’s preferred languages and scripts for the user interface, ordered by preference.