Module mas_oidc_client::types::requests

Requests and response types to interact with the OAuth 2.0 specification.

Structs

AccessTokenResponse

A successful response from the Token Endpoint.

AuthorizationCodeGrant

A request to the Token Endpoint for the Authorization Code grant type.

AuthorizationRequest

The body of a request to the Authorization Endpoint.

AuthorizationResponse

A successful response from the Authorization Endpoint.

ClientCredentialsGrant

A request to the Token Endpoint for the Client Credentials grant type.

DeviceAuthorizationRequest

A request to the Device Authorization Endpoint.

DeviceAuthorizationResponse

A successful response from the Device Authorization Endpoint.

DeviceCodeGrant

A request to the Token Endpoint for the Device Authorization grant type.

IntrospectionRequest

A request to the Introspection Endpoint.

IntrospectionResponse

A successful response from the Introspection Endpoint.

PushedAuthorizationResponse

A successful response from the Pushed Authorization Request Endpoint.

RefreshTokenGrant

A request to the Token Endpoint for refreshing an access token.

RevocationRequest

A request to the Revocation Endpoint.

Enums

AccessTokenRequest

An enum representing the possible requests to the Token Endpoint.

Display

Value that specifies how the Authorization Server displays the authentication and consent user interface pages to the End-User.

GrantType

All possible values for the grant_type parameter.

Prompt

Value that specifies whether the Authorization Server prompts the End-User for reauthentication and consent.

ResponseMode

The mechanism to be used for returning Authorization Response parameters from the Authorization Endpoint.

Constants

DEFAULT_DEVICE_AUTHORIZATION_INTERVAL_SECONDS

The default value of the interval between polling requests, if it is not set.