Function mas_oidc_client::requests::rp_initiated_logout::build_end_session_url

pub fn build_end_session_url(
    end_session_endpoint: Url,
    logout_data: LogoutData,
    rng: &mut impl Rng
) -> Result<(Url, Option<String>), Error>

Build the URL for initiating logout at the logout endpoint.

Arguments

end_session_endpoint - The URL of the issuer’s logout endpoint.

logout_data - The data necessary to build the logout request.

rng - A random number generator.

Returns

A URL to be opened in a web browser where the end-user will be able to logout of their session, and an optional state string.

The state will only be set if post_logout_redirect_uri is set. It should be present in the query when the end user is redirected to the post_logout_redirect_uri.

Errors

Returns an error if preparing the URL fails.